making parameters be double

+++ src/server/Models/NeuralTestData.cs
@@ -4,6 +4,6 @@ namespace Models
 {
     public class NeuralTestData
     {
-        public IEnumerable<List<int>> xs { get; set; }
+        public IEnumerable<List<double>> xs { get; set; }
     }
 }
